@@ -23,7 +23,7 @@ module Lokka
       set :per_page, 10
       set :admin_per_page, 200
       set :default_locale, 'en'
-      set :haml, :ugly => false, :attr_wrapper => '"'
+      set :haml, :attr_wrapper => '"', escape_html: false
       set :protect_from_csrf, true
       supported_stylesheet_templates.each do |style|
         set style, :style => :compressed
