@@ -1,5 +1,5 @@
 # config valid only for Capistrano 3.1
-lock '3.7.1'
+lock '3.9.1'
 
 set :rbenv_prefix, "RBENV_ROOT=#{fetch(:rbenv_path)} RBENV_VERSION=#{fetch(:rbenv_ruby)} rbenv exec"
 set :default_env, { 'RUBYOPT' => '--encoding=UTF-8' }
